LLM_ERP/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
transformers==4.41.2
autoawq==0.2.6
safetensors==0.4.3
Flask==3.1.0
waitress==3.0.2
cx_Oracle==8.3.0
cachetools==6.1.0
pandas==2.3.1
requests==2.32.4
# optionnels
streamlit==1.46.1
redis==6.2.0
tokenizers==0.19.1
